Fix issue when chunk ends with newline
The code would always end a message at the first chunk boundary to end
in a newline, because `String.split("a\n")` yields `["a", ""]` and the
empty string sets off the case which was supposed to detect
double-newline.

My fix feels extremely inelegant, but I think it may solve the issue.
I'd be happy to see a better approach!